+++ CHANGELOG.md
@@ -10,6 +10,9 @@
   - fix(ListOption): Fetch from correct endpoint for example
   - fix(Autocompleter): check if its mounted before modifying state
   - feat(tooltip): Initial creation / promotion of tooltip component, intended to display short help or reminder text
+  - feat(help-icon): Initial creation HelpIcon component
+  - feat(form-control): expose `tooltip` prop which renders a help icon next to the label if provided
+  - feat(input): expose `tooltip` prop which is passed to form-control
 </details>
 
 ## v0.75.0